+++ src/app/admin-module/services/data-collection.service.ts
@@ -1,71 +1,63 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Localized } from '../../shared/models/localized.type';
 
 import { DataCollectionSourceForm } from '../models/data-collection-source-form.model';
 import { DataCollectionSource } from '../models/data-collection-source.interface';
 import { DataCollection } from '../models/data-collection.interface';
 import { DataCollectionDetail } from '../models/data-collection-detail.interface';
 import { environment } from '../../../environments/environment';
+import { Localized } from '../../shared/models/localized.type';
 
 @Injectable()
 export class DataCollectionService {
-  private readonly url = environment.SERVER_API + '/data-collection';
+  private readonly pageUrl = environment.SERVER_API + '/data-collection-page';
+  private readonly sourceUrl = environment.SERVER_API + '/data-collection-source';
 
   constructor(private readonly httpClient: HttpClient) {}
 
   /** */
   public getDataCollectionPageList() {
-    const url = this.url + '/page/for-analyze';
+    const url = this.pageUrl + '/for-analyze';
     return this.httpClient.get<DataCollection[]>(url);
   }
 
   /** */
   public getDataCollectionPageDetail(id: number) {
-    const url = this.url + '/page/detail/' + id;
+    const url = this.pageUrl + '/detail/' + id;
     return this.httpClient.get<DataCollectionDetail>(url);
   }
 
   /** */
   public getDataCollectionSource() {
-    const url = this.url + '/source/list';
+    const url = this.sourceUrl + '/list';
     return this.httpClient.get<DataCollectionSource[]>(url);
   }
 
   /** */
   public getDataCollectionSourceDetail(id: number) {
-    const url = this.url + '/source/detail/' + id;
+    const url = this.sourceUrl + '/detail/' + id;
     return this.httpClient.get<DataCollectionSource>(url);
   }
 
   /**
-   * Update list of Ingredients in Recipe
+   *
    * @param id
-   * @param requiredIngredients
-   * @param optionalIngredients
+   * @param data
    */
-  public saveDataCollectionIngredients(
-    id: number,
-    requiredIngredients: number[],
-    optionalIngredients: number[]
-  ) {
-    const url = this.url + '/update/recipe-ingredients';
-    return this.httpClient.patch(url, {
-      id,
-      requiredIngredients,
-      optionalIngredients,
-    });
+  public updatePageData(id: number, data: string) {
+    const url = this.pageUrl + '/page-data';
+    return this.httpClient.put(url, {id, data})
   }
 
   /** */
   public deletePageData(id: number) {
-    const url = this.url + '/delete-data';
+    const url = this.pageUrl + '/delete-data';
     return this.httpClient.post(url, { id });
   }
 
   /** */
   public approvePageData(id: number) {
-    const url = this.url + '/approve-data';
+    const url = this.pageUrl + '/approve-data';
     return this.httpClient.post(url, { id });
   }
 
@@ -75,7 +67,7 @@ export class DataCollectionService {
    * @param link
    */
   public addSourceLink(id: number, link: string) {
-    const url = this.url + '/source/add-link';
+    const url = this.pageUrl + '/add-link';
     return this.httpClient.post(url, { id, link });
   }
 
@@ -84,7 +76,7 @@ export class DataCollectionService {
    * @param formData
    */
   public addSource(formData: DataCollectionSourceForm) {
-    const url = this.url + '/source/create';
+    const url = this.sourceUrl + '/create';
     return this.httpClient.post(url, formData);
   }
 
@@ -94,7 +86,7 @@ export class DataCollectionService {
    * @param source
    */
   public updateDataSource(id: number, source: DataCollectionSource) {
-    const url = this.url + '/update/data-source';
+    const url = this.sourceUrl + '/data-source';
     return this.httpClient.patch<DataCollectionSource>(url, { id, config: source });
   }
 
@@ -105,7 +97,7 @@ export class DataCollectionService {
    * @param locale
    */
   public updatePageTitleLocale(id: number, title: string, locale: Localized) {
-    const url = this.url + '/update/page-title-locale';
+    const url = this.pageUrl + '/page-title-locale';
     return this.httpClient.patch(url, { id, title, locale });
   }
 }