+++ src/app/data-collection/components/localization-dialog/localization-dialog.component.ts
@@ -1,0 +1,8 @@
+import { Component, ViewEncapsulation } from '@angular/core';
+
+@Component({
+  selector: 'app-localization-dialog',
+  template: ` <p>localization-dialog works!</p> `,
+  encapsulation: ViewEncapsulation.None,
+})
+export class LocalizationDialogComponent {}